backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
bcrypt==4.2.1
slowapi==0.1.9
redis==5.2.1
